2. Тесты для идеальный вес[#259438]

+++ src/ru/job4j/calculator/Fit.java
@@ -2,12 +2,12 @@ package ru.job4j.calculator;
 
 public class Fit {
 
-    public static float manWeight(float height) {
-        return (float) ((height - 100) * 1.15);
+    public static double  manWeight(double  height) {
+        return ((height - 100) * 1.15);
     }
 
-    public static float womanWeight(float height) {
-        return (float) ((height - 110) * 1.15);
+    public static double  womanWeight(double  height) {
+        return ((height - 110) * 1.15);
     }
 
     public static void main(String[] args) {
